Tests folder restructure
